Rework implementation
