--- src/Type/ActiveRecordFindBySqlReturnTypeExtension.php
@@ -1,16 +1,0 @@
-<?php
-declare(strict_types=1);
-
-namespace ErickSkrauch\PHPStan\Yii2\Type;
-
-/**
- * This extension extends the standard return type generated by PHPStan to the ActiveRecord::findBySql(),
- * so that we can further remember which model we are dealing with
- */
-final class ActiveRecordFindBySqlReturnTypeExtension extends AbstractActiveRecordFindReturnTypeExtension {
-
-    protected function methodName(): string {
-        return 'findBySql';
-    }
-
-}
