+++ src/Type/ActiveRecordFindReturnTypeExtension.php
@@ -3,14 +3,90 @@ declare(strict_types=1);
 
 namespace ErickSkrauch\PHPStan\Yii2\Type;
 
+use PhpParser\Node\Expr\StaticCall;
+use PhpParser\Node\Expr\Variable;
+use PhpParser\Node\Name;
+use PHPStan\Analyser\Scope;
+use PHPStan\Reflection\MethodReflection;
+use PHPStan\Reflection\ParametersAcceptorSelector;
+use PHPStan\Reflection\ReflectionProvider;
+use PHPStan\ShouldNotHappenException;
+use PHPStan\Type\DynamicStaticMethodReturnTypeExtension;
+use PHPStan\Type\NeverType;
+use PHPStan\Type\Type;
+use PHPStan\Type\TypeCombinator;
+use yii\db\ActiveRecord;
+use yii\db\ActiveRecordInterface;
+
 /**
  * This extension extends the standard return type generated by PHPStan to the ActiveRecord::find(),
  * so that we can further remember which model we are dealing with
  */
-final class ActiveRecordFindReturnTypeExtension extends AbstractActiveRecordFindReturnTypeExtension {
+final class ActiveRecordFindReturnTypeExtension implements DynamicStaticMethodReturnTypeExtension {
 
-    protected function methodName(): string {
-        return 'find';
+    private ReflectionProvider $reflectionProvider;
+
+    public function __construct(ReflectionProvider $reflectionProvider) {
+        $this->reflectionProvider = $reflectionProvider;
+    }
+
+    public function getClass(): string {
+        return ActiveRecordInterface::class;
+    }
+
+    public function isStaticMethodSupported(MethodReflection $methodReflection): bool {
+        return in_array($methodReflection->getName(), ['find', 'findBySql', 'findByCondition'], true);
+    }
+
+    public function getTypeFromStaticMethodCall(
+        MethodReflection $methodReflection,
+        StaticCall $methodCall,
+        Scope $scope
+    ): ?Type {
+        $calledOn = $methodCall->class;
+        $declaringClass = $methodReflection->getDeclaringClass();
+        // The implementations of the ::findBySql() and ::findByCondition() methods rely on the ::find() method,
+        // so unless they have been overridden, we return the ::find() method type
+        if ($methodReflection->getName() !== 'find' && $declaringClass->getName() === ActiveRecord::class) {
+            $findMethod = $declaringClass->getMethod('find', $scope);
+            $findCall = new StaticCall($calledOn, 'find'); // According to the Yii2 implementation, this call will have no arguments
+
+            return $this->getTypeFromStaticMethodCall($findMethod, $findCall, $scope);
+        }
+
+        if ($calledOn instanceof Name) {
+            return $this->createType($scope->resolveName($calledOn), $methodReflection->getName(), $scope);
+        }
+
+        $types = [];
+        if ($calledOn instanceof Variable) {
+            foreach ($scope->getType($calledOn)->getConstantStrings() as $constantString) {
+                if (!$constantString->isClassStringType()->yes()) {
+                    return new NeverType();
+                }
+
+                $types[] = $this->createType($constantString->getValue(), $methodReflection->getName(), $scope);
+            }
+
+            return TypeCombinator::union(...$types);
+        }
+
+        return null;
+    }
+
+    private function createType(string $modelClass, string $methodName, Scope $scope): Type {
+        $method = $this->reflectionProvider->getClass($modelClass)->getMethod($methodName, $scope);
+        $returnType = ParametersAcceptorSelector::selectSingle($method->getVariants())->getReturnType();
+        if (!$returnType->isObject()->yes()) {
+            throw new ShouldNotHappenException();
+        }
+
+        $types = [];
+        foreach ($returnType->getObjectClassNames() as $className) {
+            $types[] = new ActiveQueryObjectType(new ActiveRecordObjectType($modelClass), $className);
+        }
+
+        return TypeCombinator::union(...$types);
     }
 
 }
